requirements.txt
alembic==1.4.3
astroid==2.4.2
attrs==20.3.0
Babel==2.9.0
bandit==1.7.0
behave==1.2.6
blinker==1.4
boto3==1.16.63
botocore==1.19.63
cachelib==0.1.1
Cerberus==1.3.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
connexion==2.7.0
coverage==5.4
cryptography==3.2.1
pycryptodome==3.9.9
elasticsearch==6.4.0
Faker==5.0.2
fastavro==1.3.2
flake8==3.8.4
flake8-polyfill==1.0.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Injector==0.12.3
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-OAuthlib==0.9.6
Flask-Seeder==1.2.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.13
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
injector==0.18.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.4
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
nose2==0.10.0
oauthlib==2.1.0
openapi-spec-validator==0.2.9
parse==1.19.0
parse-type==0.5.2
pbr==5.5.1
prance==0.19.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==3.13.2
python-interface==1.6.0
python-socketio==4.6.0
pytz==2020.5
PyYAML==5.4.1
radon==4.3.2
redis==3.5.3
requests==2.25.1
requests-aws4auth==1.0.1
requests-oauthlib==1.1.0
rq==1.7.0
s3transfer==0.3.4
semver==2.13.0
sentry-sdk==0.20.2
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
stevedore==3.3.0
stringcase==1.2.0
swagger-ui-bundle==0.0.8
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==0.16.1
wrapt==1.12.1
zipp==3.4.0
gunicorn==20.0.4
